requirements.txt
streamlit
google-api-python-client
google-auth
google-auth-oauthlib
gspread
openai
beautifulsoup4